Added DataAvailableEventArgs
